feature : payment system(선결제, 결제취소, 결제유효성검사)

+++ SpaceZBE/src/main/java/com/spaceZ/serviceInfo/SpaceInfoDAO.java
@@ -9,13 +9,13 @@ import com.spaceZ.review.ReviewVO;
 @Component
 public interface SpaceInfoDAO {
 	
-	public SpaceInfoVO selectOne(int spaceId);
+	public SpaceInfoVO selectOne(long spaceId);
 
-	public double getRating(int spaceId);
+	public double getRating(long spaceId);
 
-	public List<ReplyVO> getReplys(int spaceId);
+	public List<ReplyVO> getReplys(long spaceId);
 
-	public List<ReviewVO> getReviews(int spaceId);
+	public List<ReviewVO> getReviews(long spaceId);
 
-	public List<QnaVO> getQnas(int spaceId);
+	public List<QnaVO> getQnas(long spaceId);
 }
